Fixed rare overflow bug in addition. Added a Naïve linear congruential generator to highlight Float32Exp's use case.

+++ float32exp/src/main/java/com/tbohne/util/math/Float32Exp.java
@@ -93,7 +93,7 @@ public class Float32Exp extends Float32ExpSharedBase implements Float32ExpChaine
             otherExponent = exponent;
             exponent = t;
         }
-        int diff = exponent - otherExponent;
+        long diff = ((long) exponent) - otherExponent;
         if (diff < INT_MAX_BITS) {
             long l = (((long) significand) << diff) + otherSignificand;
             setNormalized(l, ((long) exponent) - diff);
